requirements.txt
auto_mix_prep==0.2.0
matplotlib==3.9.1
numpy==2.0.1
pandas==1.5.3
plotly==5.22.0
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.14.0
